+++ src/window_main.c
@@ -24,6 +24,7 @@
 #include "window_dialogs.h"
 #include "window_rotation.h"
 #include "window_combo_box.h"
+#include "window_check_button.h"
 #include "window_auto_adjust_size.h"
 #include "window_auto_adjust_size_max_w.h"
 
@@ -52,6 +53,11 @@ static ret_t on_open_combo_box(void *ctx, event_t *e) {
   return RET_OK;
 }
 
+static ret_t on_open_check_button(void *ctx, event_t *e) {
+  window_check_button_open();
+  return RET_OK;
+}
+
 static ret_t on_open_dialogs(void *ctx, event_t *e) {
   window_dialogs_open();
   return RET_OK;
@@ -75,6 +81,7 @@ ret_t window_main_open(void) {
 
   widget_child_on(win, "rotation", EVT_CLICK, on_open_rotation, win);
   widget_child_on(win, "combo_box", EVT_CLICK, on_open_combo_box, win);
+  widget_child_on(win, "check_button", EVT_CLICK, on_open_check_button, win);
   widget_child_on(win, "dialogs", EVT_CLICK, on_open_dialogs, win);
   widget_child_on(win, "button", EVT_CLICK, on_open_button, win);